txs autopay_batch uses ending epoch
